refactor: 공공 API에서 주소도 받을 수 있도록 리팩토링 진행
[#23]
