+++ backend/src/main/java/com/carffeine/carffeine/domain/chargerStation/charger/Charger.java
@@ -38,6 +38,8 @@ public class Charger {
     // 추후에 enum으로 바꿀 수 있으면 바꾸기
     private String type;
 
+    private String address;
+
     private BigDecimal price;
 
     @Column(scale = 2)
