+++ backend/src/main/java/com/carffeine/carffeine/service/chargerStation/dto/ChargeStationInfoRequest.java
@@ -73,6 +73,7 @@ public record ChargeStationInfoRequest(
                         .stationId(statId)
                         .chargerId(chgerId)
                         .type(chgerType)
+                        .address(addr)
                         .chargerStatus(toChargerStatus())
                         .capacity(parseBigDecimalFromString(output))
                         .method(method)
