+++ backend/src/test/java/com/carffeine/carffeine/fixture/chargerStation/ChargerFixture.java
@@ -15,6 +15,7 @@ public class ChargerFixture {
             .chargerId("01")
             .price(BigDecimal.valueOf(324.4))
             .method("단독")
+            .address("서울특별시 강남구 선릉로100길 2 선릉역1")
             .capacity(BigDecimal.valueOf(50.0))
             .type("급속")
             .chargerStatus(ChargerStatus.builder()
@@ -30,6 +31,7 @@ public class ChargerFixture {
             .chargerId("02")
             .price(BigDecimal.valueOf(324.4))
             .method("단독")
+            .address("서울특별시 강남구 선릉로100길 2 선릉역2")
             .capacity(BigDecimal.valueOf(50.0))
             .type("급속")
             .chargerStatus(ChargerStatus.builder()
